Fixed some warnings realted to implicit casts
Fixed some warnings realted to implicit casts, checked if those casts
were intended, and used static_cast if they were.
Fixed them if they were not.

+++ include/song.hpp
@@ -43,8 +43,8 @@ public:
 
     // TODO consider making these strings as well
     std::uint32_t m_audio_start = 0;
-    std::uint32_t m_duration = 0;
-    std::uint32_t m_delay = 0;
+    std::uint64_t m_duration = 0;
+    std::uint64_t m_delay = 0;
 
     std::uint64_t m_play_counter = 0;
